+++ CHANGELOG.md
@@ -2,6 +2,47 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [8.40.5](https://github.com/GetStream/stream-chat-js/compare/v8.40.4...v8.40.5) (2024-09-11)
+
+
+### Bug Fixes
+
+* adding message to an empty message list ([#1358](https://github.com/GetStream/stream-chat-js/issues/1358)) ([599385d](https://github.com/GetStream/stream-chat-js/commit/599385d97b97e992433b04c00f63266f055b3099))
+* allow hydrating thread's channel with an empty members list ([#1359](https://github.com/GetStream/stream-chat-js/issues/1359)) ([0946f45](https://github.com/GetStream/stream-chat-js/commit/0946f450000154e259d9340e0f6a861a432b1764))
+
+### [8.40.4](https://github.com/GetStream/stream-chat-js/compare/v8.40.3...v8.40.4) (2024-09-11)
+
+
+### Bug Fixes
+
+* adding message to an empty message list ([#1356](https://github.com/GetStream/stream-chat-js/issues/1356)) ([acd55ca](https://github.com/GetStream/stream-chat-js/commit/acd55ca812b56dd551d749336386affb7f66732c))
+
+### [8.40.3](https://github.com/GetStream/stream-chat-js/compare/v8.40.2...v8.40.3) (2024-09-10)
+
+
+### Features
+
+* add partialUpdateMember endpoint ([#1344](https://github.com/GetStream/stream-chat-js/issues/1344)) ([b9b0393](https://github.com/GetStream/stream-chat-js/commit/b9b03935baf5e327d47615fade8f52f6046d0bbf))
+
+
+### Bug Fixes
+
+* hydrate channel instance from thread response ([#1354](https://github.com/GetStream/stream-chat-js/issues/1354)) ([35abae1](https://github.com/GetStream/stream-chat-js/commit/35abae1bac45f6abc5ebceb6a47b34abf2c7329e))
+
+### [8.40.2](https://github.com/GetStream/stream-chat-js/compare/v8.40.1...v8.40.2) (2024-09-06)
+
+
+### Bug Fixes
+
+* **threads:** restore reply soft-deletion ([#1352](https://github.com/GetStream/stream-chat-js/issues/1352)) ([7204830](https://github.com/GetStream/stream-chat-js/commit/72048305e7e6795d23d4b5fae4a1fa96c4c9cd38))
+
+### [8.40.1](https://github.com/GetStream/stream-chat-js/compare/v8.40.0...v8.40.1) (2024-09-04)
+
+
+### Bug Fixes
+
+* include system messages in unread count ([#1350](https://github.com/GetStream/stream-chat-js/issues/1350)) ([9f1d27a](https://github.com/GetStream/stream-chat-js/commit/9f1d27a0c5f652b625fc73bdd26aaf2d1189c317))
+
 ## [8.40.0](https://github.com/GetStream/stream-chat-js/compare/v8.39.0...v8.40.0) (2024-09-02)